app/requirements.txt
Django==4.2.1
djangorestframework==3.14.0
djoser==2.1.0
drf-yasg==1.21.4
gunicorn==20.1.0
python-dotenv==0.21.0